final_pjt/final_pjt_semi_final/back-server/requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
Faker==18.4.0
idna==3.4
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.6
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
toposort==1.10
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
